+++ src/blueprints/telegram_bot/webhook/commands/__init__.py
@@ -1,3 +1,27 @@
+"""
+How to create handler for the dispatcher:
+- you should provide one function. This function
+will be called to handle an incoming Telegram message
+- dispatcher handles any handler exceptions, so, if
+any unexpected error occurs you can raise the exception.
+Sure, you can raise your own exception for nice debug
+- edit dispatcher module in order to register your handler
+- handler should accept only `(*args, **kwargs)` arguments
+- dispatcher passes in `**kwargs`: `route_source: RouteSource`,
+`message_events: Set[str]` (`str` it is `DispatcherEvent` values),
+`user_id: int`, `chat_id: int`, `message: TelegramMessage`.
+These values can have `None` value, so, check for it before using.
+For another values in `*args` and `**kwargs` see documentation of
+functions from call stack of dispatcher function.
+
+How to create stateful chat:
+- use `set_disposable_handler`, `subscribe_handler`,
+`unsubcribe_handler`, `set/get/delete_user/user_chat/chat_data`
+- if you want to provide Enum, then provide value of that enum,
+not object directly
+"""
+
+
 from .common.names import CommandName
 from .unknown import handle as unknown_handler
 from .help import handle as help_handler