+++ src/blueprints/telegram_bot/webhook/commands/about.py
@@ -4,7 +4,7 @@ from src.api import telegram
 from src.blueprints.utils import absolute_url_for
 
 
-def handle():
+def handle(*args, **kwargs):
     """
     Handles `/about` command.
     """