+++ src/blueprints/telegram_bot/webhook/commands/settings.py
@@ -12,7 +12,7 @@ from .common.responses import (
 
 @register_guest
 @get_db_data
-def handle():
+def handle(*args, **kwargs):
     """
     Handles `/settings` command.
     """