+++ src/blueprints/telegram_bot/webhook/commands/unknown.py
@@ -4,7 +4,7 @@ from src.api import telegram
 from . import CommandName
 
 
-def handle():
+def handle(*args, **kwargs):
     """
     Handles unknown command.
     """