+++ src/blueprints/telegram_bot/webhook/commands/yd_revoke.py
@@ -9,7 +9,7 @@ from . import CommandName
 
 
 @get_db_data
-def handle():
+def handle(*args, **kwargs):
     """
     Handles `/yandex_disk_revoke` command.