+++ src/blueprints/telegram_bot/webhook/dispatcher.py
@@ -1,41 +1,135 @@
-from typing import Union, Callable
+from typing import (
+    Union,
+    Callable,
+    Set
+)
+from collections import deque
 
+from src.extensions import redis_client
 from . import commands
 from .commands import CommandName
+from .dispatcher_events import (
+    DispatcherEvent,
+    RouteSource
+)
 from .telegram_interface import (
     Message as TelegramMessage
 )
+from .stateful_chat import (
+    get_disposable_handler,
+    delete_disposable_handler,
+    get_subscribed_handlers
+)
 
 
-def dispatch(message: TelegramMessage) -> Callable:
+def intellectual_dispatch(
+    message: TelegramMessage
+) -> Callable:
     """
-    Dispatch to handler of a message.
-    It handles message by different ways: tries to
-    read the content, tries to guess the command,
-    tries to implement stateful dialog, tries to pass
-    most appropriate arguments, and so on.
-    So, most appropriate handler will be returned
-    for incoming message.
+    Intellectual dispatch to handlers of a message.
+    Provides support for stateful chat (if Redis is enabled).
+
+    Priority of handlers:
+    1) if disposable handler exists and message events matches,
+    then only that handler will be called and after removed
+    2) if subscribed handlers exists, then only ones with events
+    matched to message events will be called. If nothing is matched,
+    then forwarding to № 3
+    3) attempt to get first `bot_command` entity from message.
+    If nothing found, then forwarding to № 4
+    4) guessing of command that user assumed based on
+    content of message
+
+    Events matching:
+    - if at least one event matched, then that handler will be
+    marked as "matched".
+
+    If stateful chat not enabled, then № 1 and № 2 will be skipped.
+
+    Note: there can be multiple handlers picked for single message.
+    Order of execution not determined.
 
     :param message:
     Incoming Telegram message.
 
     :returns:
-    It is guaranteed that most appropriate callable
-    handler will be returned. It is a handler for
-    incoming message with already configured arguments,
-    and you should call this with no arguments
-    (but you can pass any if you want).
+    It is guaranteed that most appropriate callable handler that
+    not raises an error will be returned. Function arguments already
+    configured, but you can also provided your own through `*args`
+    and `**kwargs`. You should call this function in order to run
+    handlers (there can be multiple handlers in one return function).
     """
-    command = message.get_entity_value("bot_command")
+    user_id = message.get_user().id
+    chat_id = message.get_chat().id
+    stateful_chat_is_enabled = redis_client.is_enabled
+    disposable_handler = None
+    subscribed_handlers = None
 
-    if command is None:
-        command = guess_bot_command(message)
+    if stateful_chat_is_enabled:
+        disposable_handler = get_disposable_handler(user_id, chat_id)
+        subscribed_handlers = get_subscribed_handlers(user_id, chat_id)
 
-    handler = direct_dispatch(command)
+    message_events = (
+        detect_events(message) if (
+            disposable_handler or
+            subscribed_handlers
+        ) else None
+    )
+    handler_names = deque()
+    route_source = None
+
+    if disposable_handler:
+        match = match_events(
+            message_events,
+            disposable_handler["events"]
+        )
+
+        if match:
+            route_source = RouteSource.DISPOSABLE_HANDLER
+            handler_names.append(disposable_handler["name"])
+            delete_disposable_handler(user_id, chat_id)
+
+    if (
+        subscribed_handlers and
+        not handler_names
+    ):
+        for handler in subscribed_handlers:
+            match = match_events(
+                message_events,
+                handler["events"]
+            )
+
+            if match:
+                route_source = RouteSource.SUBSCRIBED_HANDLER
+                handler_names.append(handler["name"])
+
+    if not handler_names:
+        command = message.get_entity_value("bot_command")
+
+        if command:
+            route_source = RouteSource.DIRECT_COMMAND
+            handler_names.append(command)
+
+    if not handler_names:
+        route_source = RouteSource.GUESSED_COMMAND
+        handler_names.append(guess_bot_command(message))
 
     def method(*args, **kwargs):
-        handler(*args, **kwargs)
+        for handler_name in handler_names:
+            handler_method = direct_dispatch(handler_name)
+
+            try:
+                handler_method(
+                    *args,
+                    **kwargs,
+                    user_id=user_id,
+                    chat_id=chat_id,
+                    message=message,
+                    route_source=route_source,
+                    message_events=message_events
+                )
+            except Exception as error:
+                print(handler_name, error)
 
     return method
 
@@ -91,7 +185,7 @@ def direct_dispatch(
 def guess_bot_command(
     message: TelegramMessage,
     fallback: CommandName = CommandName.HELP
-) -> CommandName:
+) -> str:
     """
     Tries to guess which bot command user
     assumed based on content of a message.
@@ -100,7 +194,7 @@ def guess_bot_command(
     Fallback command which will be returned if unable to guess.
 
     :returns:
-    Guessed bot command based on a message.
+    Guessed bot command name based on a message.
     """
     command = fallback
     raw_data = message.raw_data
@@ -118,4 +212,76 @@ def guess_bot_command(
     elif (message.get_entity_value("url") is not None):
         command = CommandName.UPLOAD_URL
 
-    return command
+    return command.value
+
+
+def detect_events(
+    message: TelegramMessage
+) -> Set[str]:
+    """
+    :returns:
+    Detected dispatcher events.
+    See `DispatcherEvent` documentation for more.
+    Note: it is strings values, because these values
+    should be compared with Redis values, which is
+    also strings.
+    """
+    events = set()
+    entities = message.get_entities()
+    photo, document, audio, video, voice = map(
+        lambda x: x in message.raw_data,
+        ("photo", "document", "audio", "video", "voice")
+    )
+    url, hashtag, email, bot_command = map(
+        lambda x: any(e.type == x for e in entities),
+        ("url", "hashtag", "email", "bot_command")
+    )
+    plain_text = message.get_plain_text()
+
+    if photo:
+        events.add(DispatcherEvent.PHOTO.value)
+
+    if document:
+        events.add(DispatcherEvent.FILE.value)
+
+    if audio:
+        events.add(DispatcherEvent.AUDIO.value)
+
+    if video:
+        events.add(DispatcherEvent.VIDEO.value)
+
+    if voice:
+        events.add(DispatcherEvent.VOICE.value)
+
+    if url:
+        events.add(DispatcherEvent.URL.value)
+
+    if hashtag:
+        events.add(DispatcherEvent.HASHTAG.value)
+
+    if email:
+        events.add(DispatcherEvent.EMAIL.value)
+
+    if bot_command:
+        events.add(DispatcherEvent.BOT_COMMAND.value)
+
+    if plain_text:
+        events.add(DispatcherEvent.PLAIN_TEXT.value)
+
+    if not len(events):
+        events.add(DispatcherEvent.NONE.value)
+
+    return events
+
+
+def match_events(
+    a: Set[str],
+    b: Set[str]
+) -> bool:
+    """
+    Checks if two groups of events are matched.
+
+    :returns:
+    `True` - match found, `False` otherwise.
+    """
+    return any(x in b for x in a)